refactor(cart): split cart.actions into single-responsibility modules
Separated cart.actions.ts into smaller, single-responsibility files:

- Moved addItemToCart to add-item-to-cart.action.ts
- Moved removeItemFromCart to remove-item-from-cart.action.ts
- Moved getCart to get-cart.action.ts
- Extracted getSessionCartIdAndUserId to session-cart.util.ts

Improves code readability, testability, and maintainability by following the separation of concerns principle.
