+++ app/(shop)/produto/[slug]/detalhes/page.tsx
@@ -4,7 +4,7 @@ import { productRepositoryAdapter } from "@/infra/adapters/product/product-repos
 import { ProductDetails } from "@/components/product/product-details";
 import { ProductPurchaseCard } from "@/components/product/product-purchase-card";
 import { ProductImages } from "@/components/product/product-images";
-import { getCart } from "@/lib/actions/cart.actions";
+import { getCart } from "@/lib/actions/cart.actions/get-cart.action.ts";
 import { initialActionDataState } from "@/lib/actions/utils.actions";
 import { newCartEntity } from "@/domain/entities/cart.entity";
