+++ containers/cart/cart-item-action-container.tsx
@@ -3,7 +3,8 @@ import { useRouter } from "next/navigation";
 import { toast } from "sonner";
 
 import { CartItemEntity } from "@/domain/entities/cart-item.entity";
-import { addItemToCart, removeItemFromCart } from "@/lib/actions/cart.actions";
+import { addItemToCart } from "@/lib/actions/cart.actions/add-item-to-cart.action";
+import { removeItemFromCart } from "@/lib/actions/cart.actions/remove-item-from-cart.action";
 import { initialActionState } from "@/lib/actions/utils.actions";
 import { toastSuccess } from "@/components/common/toast-success";
 import { CART_VIEW_PATH } from "@/lib/constants/routes";
