+++ lib/actions/cart.actions/session-cart.util.ts
@@ -1,0 +1,24 @@
+"use server";
+import { cookies } from "next/headers";
+
+import { auth } from "@/auth";
+
+export async function getSessionCartIdAndUserId(): Promise<{
+  sessionCartId: string;
+  userId?: string;
+}> {
+  // Check for cart cookie
+  const sessionCartId = (await cookies()).get("sessionCartId")?.value;
+  if (!sessionCartId) {
+    throw new Error("Não foi possível encontrar o ID do carrinho na sessão");
+  }
+
+  // Get session and user ID
+  const session = await auth();
+  const userId = session?.user?.id;
+
+  return {
+    sessionCartId,
+    userId,
+  };
+}
